+++ chat/common.ts
@@ -46,6 +46,11 @@ export class Settings implements ISettings {
     risingAdScore = true;
     risingLinkPreview = true;
     risingAutoCompareKinks = true;
+
+    risingAutoExpandCustomKinks = true;
+    risingCharacterPreview = true;
+    risingComparisonInUserMenu = true;
+    risingComparisonInSearch = true;
 }